Make wxGUIAppTraits::GetDesktopEnvironment() accessible to wxQt port under *Nix
