+++ src/gtk/utilsgtk.cpp
@@ -19,7 +19,6 @@
 
 #include "wx/apptrait.h"
 #include "wx/process.h"
-#include "wx/sysopt.h"
 #include "wx/vector.h"
 
 #include "wx/gtk/private/timer.h"
@@ -42,12 +41,6 @@
 #include <sys/types.h>
 #ifdef __UNIX__
 #include <unistd.h>
-#endif
-
-#if wxUSE_DETECT_SM
-    #include <X11/SM/SMlib.h>
-
-    #include "wx/unix/utilsx11.h"
 #endif
 
 GdkWindow* wxGetTopLevelGDK();
@@ -173,42 +166,6 @@ wxTimerImpl *wxGUIAppTraits::CreateTimerImpl(wxTimer *timer)
 }
 
 #endif // wxUSE_TIMER
-
-#if wxUSE_DETECT_SM
-static wxString GetSM()
-{
-    wxX11Display dpy;
-    if ( !dpy )
-        return wxEmptyString;
-
-    char smerr[256];
-    char *client_id;
-    SmcConn smc_conn = SmcOpenConnection(nullptr, nullptr,
-                                         999, 999,
-                                         0 /* mask */, nullptr /* callbacks */,
-                                         nullptr, &client_id,
-                                         WXSIZEOF(smerr), smerr);
-
-    if ( !smc_conn )
-    {
-        // Don't report error if there is no session manager at all
-        if (getenv("SESSION_MANAGER"))
-        {
-            wxLogDebug("Failed to connect to session manager: %s", smerr);
-        }
-        return wxEmptyString;
-    }
-
-    char *vendor = SmcVendor(smc_conn);
-    wxString ret = wxString::FromAscii( vendor );
-    free(vendor);
-
-    SmcCloseConnection(smc_conn, 0, nullptr);
-    free(client_id);
-
-    return ret;
-}
-#endif // wxUSE_DETECT_SM
 
 
 //-----------------------------------------------------------------------------
@@ -368,43 +325,6 @@ bool wxGUIAppTraits::ShowAssertDialog(const wxString& msg)
 #endif // wxDEBUG_LEVEL
 
     return wxAppTraitsBase::ShowAssertDialog(msg);
-}
-
-#endif // __UNIX__
-
-#if defined(__UNIX__)
-
-wxString wxGUIAppTraits::GetDesktopEnvironment() const
-{
-    wxString de = wxSystemOptions::GetOption(wxT("gtk.desktop"));
-    if (!de.empty())
-        return de;
-
-    de = wxGetenv(wxS("XDG_CURRENT_DESKTOP"));
-    if (!de.empty())
-    {
-        // Can be a colon separated list according to
-        // https://wiki.archlinux.org/title/Environment_variables#Examples
-        de = de.BeforeFirst(':');
-    }
-#if wxUSE_DETECT_SM
-    if ( de.empty() )
-    {
-        static const wxString s_SM(GetSM());
-        de = s_SM;
-        de.Replace(wxS("-session"), wxString());
-    }
-#endif // wxUSE_DETECT_SM
-
-    de.MakeUpper();
-    if (de.Contains(wxS("GNOME")))
-        de = wxS("GNOME");
-    else if (de.Contains(wxS("KDE")))
-        de = wxS("KDE");
-    else if (de.Contains(wxS("XFCE")))
-        de = wxS("XFCE");
-
-    return de;
 }
 
 #endif // __UNIX__
